+++ oap-server/analyzer/log-analyzer/src/main/java/org/apache/skywalking/oap/log/analyzer/provider/log/listener/TrafficAnalysisListener.java
@@ -84,7 +84,6 @@ public class TrafficAnalysisListener implements LogAnalysisListener {
             instanceMeta.setServiceId(serviceId);
             instanceMeta.setName(namingControl.formatInstanceName(logData.getServiceInstance()));
             instanceMeta.setTimeBucket(timeBucket);
-            instanceMeta.setLayer(layer);
 
         }
         // to endpoint traffic