+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/query/IoTDBMetadataQueryDAO.java
@@ -268,7 +268,6 @@ public class IoTDBMetadataQueryDAO implements IMetadataQueryDAO {
             serviceInstance.setId(instanceTraffic.id());
             serviceInstance.setName(instanceTraffic.getName());
             serviceInstance.setInstanceUUID(serviceInstance.getId());
-            serviceInstance.setLayer(instanceTraffic.getLayer().name());
 
             JsonObject properties = instanceTraffic.getProperties();
             if (properties != null) {